feat: migrate from PostgreSQL to embedded SQLite (libsql)
Drops the Railway/Postgres dependency in favor of @libsql/client, so the
app runs with zero external setup (matches the pattern already used in
mercado-pro-desktop). Schema, raw-SQL bootstrap, and db client rewritten
for SQLite; the ~9 call sites with Postgres-specific ::int casts and
filter(where...) syntax updated for SQLite compatibility. Adds
output: "standalone" to next.config.ts in preparation for Electron
packaging.

Also fixes two false-positive AI build diagnoses in engine.ts (matching
on a bare filename/word instead of the actual error text — same class of
bug as the earlier "app web" false positive), and adds automatic Gradle
cache corruption recovery to the worker script: it now detects
CorruptedCacheException and clears + retries once instead of requiring
a manual cache wipe every time.

+++ src/app/api/builds/status/route.ts
@@ -26,7 +26,7 @@ export async function GET() {
       framework: projects.framework,
       durationMs: builds.durationMs,
       createdAt: builds.createdAt,
-      artifactCount: sql<number>`(select count(*)::int from ${artifacts} where ${artifacts.buildId} = ${builds.id})`,
+      artifactCount: sql<number>`(select count(*) from ${artifacts} where ${artifacts.buildId} = ${builds.id})`,
     })
     .from(builds)
     .innerJoin(projects, eq(builds.projectId, projects.id))
